pincart/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
playwright==1.49.1
openai==1.58.1
supabase==2.11.0
stripe==11.4.1
python-dotenv==1.0.1
httpx==0.28.1
python-multipart==0.0.20
celery[redis]==5.3.4
redis==5.0.1
sentry-sdk[fastapi]==1.40.0
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
